+++ src/components/side-effects/CommandInterface/index.tsx
@@ -2,8 +2,9 @@ import { useRouter } from 'next/router'
 import React, { FormEvent, RefObject, useEffect, useLayoutEffect, useRef, useState } from 'react'
 import { MdArrowForward } from 'react-icons/md'
 import { Project, useTrendingProjectsQuery } from '@/graphql/generated/gql'
-import { defaultFilters, defaultSort } from '@/components/page/overview/types'
+import { defaultSort } from '@/components/page/overview/types'
 import emailTemplate from '@/util/emailTemplate'
+import defaultFilters from '@/components/page/overview/defaultFilters'
 import defaultList from './DefaultRecommendationList'
 import CommandInterfaceOptions from './CommandInterfaceOptions'
 import CommandInterfaceModal from './CommandInterfaceModal'
